+++ src/VaultSharp/Backends/Secret/Models/MongoDb/MongoDbRoleDefinition.cs
@@ -17,13 +17,13 @@ namespace VaultSharp.Backends.Secret.Models.MongoDb
         [JsonProperty("db")]
         public string Database { get; set; }
 
-        // raja todo: make this more friendly, and string should be the output type.
-
         /// <summary>
         /// <para>[required]</para>
         /// Gets or sets the MongoDb roles. You specify both built-in roles and user-defined roles 
         /// for both the database the user is created in and for other databases.
         /// e.g. Roles = JsonConvert.SerializeObject(new object[] { "readWrite", new { role = "read", db = "bar" } }) 
+        /// Vault will create a user with the 'readWrite' built-in role on the requested database 
+        /// and the read built-in role on the 'bar' database.
         /// </summary>
         /// <value>
         /// The roles.